+++ lib/core/di/dependency_injection.dart
@@ -1,0 +1,7 @@
+import 'package:get_it/get_it.dart';
+
+final getIt = GetIt.instance;
+
+Future<void> setupGetIt() async {
+
+}
