+++ lib/core/network/api_constants.dart
@@ -1,0 +1,8 @@
+class ApiConstants {
+
+
+}
+
+class ApiErrors {
+  
+}
