+++ lib/core/network/api_service.dart
@@ -1,0 +1,3 @@
+abstract class ApiService {
+
+}
