+++ lib/features/auth/forget_password/data/model/forget_password_model.dart
@@ -1,3 +1,3 @@
-class ChatbotModel{
+class ForgetPasswordModel{
 
 }
