+++ lib/features/auth/forget_password/logic/forget_password_cubit/forget_password_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
-part 'chatbot_state.dart';
+part 'forget_password_state.dart';
 
-class ChatbotCubit extends Cubit<ChatbotState> {
-  ChatbotCubit() : super(ChatbotInitial());
+class ForgetPasswordCubit extends Cubit<ForgetPasswordState> {
+  ForgetPasswordCubit() : super(ForgetPasswordInitial());
 }
