+++ lib/features/auth/forget_password/ui/forget_password_screen.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:tranquilo_app/features/chatbot/ui/widgets/chatbot_screen_body.dart';
+import 'package:tranquilo_app/features/auth/forget_password/ui/widgets/forget_password_screen_body.dart';
 
-class ChatbotScreen extends StatelessWidget{
-  const ChatbotScreen({super.key});
+class ForgetPasswordScreen extends StatelessWidget{
+  const ForgetPasswordScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
-      body: ChatbotScreenBody(),
+      body: ForgetPasswordScreenBody(),
     );
   }
 }
