+++ lib/features/auth/forget_password/ui/widgets/forget_password_screen_body.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class ChatbotScreenBody extends StatelessWidget{
-  const ChatbotScreenBody({super.key});
+class ForgetPasswordScreenBody extends StatelessWidget{
+  const ForgetPasswordScreenBody({super.key});
 
   @override
   Widget build(BuildContext context) {
