+++ lib/features/auth/login/data/model/login_model.dart
@@ -1,3 +1,3 @@
-class ForgetPasswordModel{
+class LoginModel{
 
 }
