+++ lib/features/auth/login/logic/login_cubit/login_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
-part 'forget_password_state.dart';
+part 'login_state.dart';
 
-class ForgetPasswordCubit extends Cubit<ForgetPasswordState> {
-  ForgetPasswordCubit() : super(ForgetPasswordInitial());
+class LoginCubit extends Cubit<LoginState> {
+  LoginCubit() : super(LoginInitial());
 }
