+++ lib/features/auth/login/ui/widgets/login_screen_body.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class ForgetPasswordScreenBody extends StatelessWidget{
-  const ForgetPasswordScreenBody({super.key});
+class LoginScreenBody extends StatelessWidget{
+  const LoginScreenBody({super.key});
 
   @override
   Widget build(BuildContext context) {
