+++ lib/features/auth/sign_up/data/model/sign_up_model.dart
@@ -1,3 +1,3 @@
-class LoginModel{
+class SignUpModel{
 
 }
