+++ lib/features/auth/sign_up/data/repo/sign_up_repo.dart
@@ -1,3 +1,3 @@
-class LoginRepo{
+class SignUpRepo{
 
 }
