+++ lib/features/auth/sign_up/ui/widgets/sign_up_screen_body.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class LoginScreenBody extends StatelessWidget{
-  const LoginScreenBody({super.key});
+class SignUpScreenBody extends StatelessWidget{
+  const SignUpScreenBody({super.key});
 
   @override
   Widget build(BuildContext context) {
