+++ lib/features/chatbot/data/model/chatbot_model.dart
@@ -1,3 +1,3 @@
-class DashboardModel{
+class ChatbotModel{
 
 }
