+++ lib/features/chatbot/data/repo/chatbot_repo.dart
@@ -1,3 +1,3 @@
-class CommunityRepo{
+class ChatbotRepo{
 
 }
