+++ lib/features/chatbot/ui/chatbot_screen.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:tranquilo_app/features/community/ui/widgets/dashboard_screen_body.dart';
+import 'package:tranquilo_app/features/chatbot/ui/widgets/chatbot_screen_body.dart';
 
-class CommunityScreen extends StatelessWidget{
-  const CommunityScreen({super.key});
+class ChatbotScreen extends StatelessWidget{
+  const ChatbotScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
-      body: CommunityScreenBody(),
+      body: ChatbotScreenBody(),
     );
   }
 }
