+++ lib/features/chatbot/ui/widgets/chatbot_screen_body.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class CommunityScreenBody extends StatelessWidget{
-  const CommunityScreenBody({super.key});
+class ChatbotScreenBody extends StatelessWidget{
+  const ChatbotScreenBody({super.key});
 
   @override
   Widget build(BuildContext context) {
