+++ lib/features/community/data/model/community_model.dart
@@ -1,3 +1,3 @@
-class HomeModel{
+class CommunityModel{
 
 }
