+++ lib/features/community/data/repo/community_repo.dart
@@ -1,3 +1,3 @@
-class DashboardRepo{
+class CommunityRepo{
 
 }
