+++ lib/features/dashboard/data/model/dashboard_model.dart
@@ -1,3 +1,3 @@
-class HomeModel{
+class DashboardModel{
 
 }
