+++ lib/features/dashboard/data/repo/dashboard_repo.dart
@@ -1,3 +1,3 @@
-class HomeRepo{
+class DashboardRepo{
 
 }
