+++ lib/features/dashboard/ui/widgets/dashboard_screen_body.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class HomeScreenBody extends StatelessWidget{
-  const HomeScreenBody({super.key});
+class DashboardScreenBody extends StatelessWidget{
+  const DashboardScreenBody({super.key});
 
   @override
   Widget build(BuildContext context) {
