+++ lib/features/home/data/model/home_model.dart
@@ -1,0 +1,3 @@
+class HomeModel{
+
+}
