+++ lib/features/home/logic/home_cubit/home_state.dart
@@ -1,6 +1,5 @@
 part of 'home_cubit.dart';
 
-@immutable
 sealed class HomeState {}
 
 final class HomeInitial extends HomeState {}
