+++ lib/features/home/ui/widgets/home_screen_body.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class HomeScreenBody extends StatelessWidget{
+  const HomeScreenBody({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Column();
+  }
+}
