+++ lib/features/survey/data/model/survey_model.dart
@@ -1,3 +1,3 @@
-class HomeModel{
+class SurveyModel{
 
 }
