+++ lib/tranquilo.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class TranquiloApp extends StatelessWidget{
+  const TranquiloApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp();
+  }
+}
