+++ pubspec.yaml
@@ -37,35 +37,35 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.6
   #A build system for Dart code generation and modular compilation
-  build_runner: ^2.4.7
+  build_runner: ^2.4.12
   # State management
-  flutter_bloc: ^8.1.3
+  flutter_bloc: ^8.1.6
   # Code generator (mainly used with cubit & retrofit)
-  freezed: ^2.4.5
-  freezed_annotation: ^2.4.1
+  freezed: ^2.5.7
+  freezed_annotation: ^2.4.4
   # for dependency injection
-  get_it: ^7.6.4
+  get_it: ^7.7.0
   # Automatically generate code for converting to and from JSON by annotating Dart classes
-  json_serializable: ^6.7.1
-  json_annotation: ^4.8.1
+  json_serializable: ^6.8.0
+  json_annotation: ^4.9.0
   # Networking
-  dio: ^5.3.3
-  retrofit: ^4.0.3
-  retrofit_generator: ^8.0.4
-  pretty_dio_logger: ^1.3.1
+  dio: ^5.7.0
+  retrofit: ^4.4.1
+  retrofit_generator: ^9.1.2
+  pretty_dio_logger: ^1.4.0
   # localization
   intl: ^0.19.0
-  easy_localization: ^3.0.3
+  easy_localization: ^3.0.7
   # for splash screen
-  flutter_native_splash: ^2.3.5
+  flutter_native_splash: ^2.4.1
   # An SVG rendering and widget library for Flutter, which allows painting
-  flutter_svg: ^2.0.9
+  flutter_svg: ^2.0.10+1
   # for adapting screen and font size.
-  flutter_screenutil: ^5.9.0
+  flutter_screenutil: ^5.9.3
   # Flutter library to load and cache network images.
-  cached_network_image: ^3.3.1
+  cached_network_image: ^3.4.1
   # reading and writing simple key-value pairs
-  shared_preferences: ^2.2.3
+  shared_preferences: ^2.3.2
   # provides API to store data in secure storage.
   flutter_secure_storage: ^9.2.2
   # easy way to add shimmer effect
